renamed newtooltip to tooltip
