+++ packages/circuit-ui/components/Checkbox/Checkbox.spec.tsx
@@ -31,6 +31,13 @@ const defaultProps = {
 };
 
 describe('Checkbox', () => {
+  // ResizeObserver Mock
+  global.ResizeObserver = jest.fn().mockImplementation(() => ({
+    observe: jest.fn(),
+    unobserve: jest.fn(),
+    disconnect: jest.fn(),
+  }));
+
   /**
    * Style tests.
    */
