--- packages/circuit-ui/components/NewTooltip/NewTooltip.docs.mdx
@@ -1,39 +1,0 @@
-import { Status, Props, Story } from '../../../../.storybook/components';
-
-# Tooltip
-
-<Status.Stable />
-
-Tooltips display additional information on floating labels.
-They can be activated when users hover, focus, tap, or click.
-
-<Story id="components-newtooltip--base" />
-
-<Props />
-
-## When to use it
-
-You can use it as primary label, to associate one element with another (Icon→Action), or
-as an auxiliary description to provide a supplementary or clarifying description.
-
-## Usage guidelines
-
-- **Do** use delay before the tooltip is hidden on mouse leave.
-- **Do** provide useful, additional information or clarification.
-- **Do** use sentence case.
-- **Do** be brief.
-- **Do** use it to describe an action/icon.
-- **Do not** use delay when the tooltip is shown on mouse enter
-- **Do not** use it to communicate important information (form title attributes or notifications)
-- **Do not** put links or buttons inside of a tooltip
-- **Do not** use it too often. If you are building not self-explanatory software, try different approaches.
-
-## Variations
-
-<Story id="components-newtooltip--top" />
-
-<Story id="components-newtooltip--bottom" />
-
-<Story id="components-newtooltip--left" />
-
-<Story id="components-newtooltip--right" />
