+++ packages/circuit-ui/components/Tooltip/Tooltip.docs.mdx
@@ -30,8 +30,10 @@ as an auxiliary description to provide a supplementary or clarifying description
 
 ## Variations
 
-<Story id="components-tooltip--top-left" />
+<Story id="components-tooltip--top" />
 
-<Story id="components-tooltip--bottom-right" />
+<Story id="components-tooltip--bottom" />
 
-<Story id="components-tooltip--left-center" />
+<Story id="components-tooltip--left" />
+
+<Story id="components-tooltip--right" />
